Move internal/metrics to daemon/internal/metrics
