+++ daemon/info.go
@@ -20,9 +20,9 @@ import (
 	"github.com/docker/docker/daemon/command/debug"
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/internal/filedescriptors"
+	"github.com/docker/docker/daemon/internal/metrics"
 	"github.com/docker/docker/daemon/logger"
 	"github.com/docker/docker/dockerversion"
-	"github.com/docker/docker/internal/metrics"
 	"github.com/docker/docker/internal/platform"
 	"github.com/docker/docker/pkg/meminfo"
 	"github.com/docker/docker/pkg/parsers/kernel"
