+++ daemon/internal/metrics/plugin_unix.go
