+++ daemon/start.go
@@ -11,8 +11,8 @@ import (
 	"github.com/docker/docker/api/types/events"
 	"github.com/docker/docker/container"
 	mobyc8dstore "github.com/docker/docker/daemon/containerd"
+	"github.com/docker/docker/daemon/internal/metrics"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/internal/metrics"
 	"github.com/docker/docker/internal/otelutil"
 	"github.com/docker/docker/libcontainerd"
 	"github.com/pkg/errors"
